+++ Chapter3/BuildRestfulApiWithExpress/router/userRouter.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const userRouter = express.Router();
+
+//Handling user:
+app.get("/user/info");
+app.post("/user/info");
+
+module.exports = userRouter;
